Push modified documents
