+++ output-data/GSE/Artemis/Artemis_2011-12-28_16-20.csv
@@ -8,8 +8,8 @@
 6,2011-12-28 16:26:00,-2.8767455,4.311607,-8.202203,-273.5818583170573,3.691596031188965,13.748967488606771,26.06695493062337,303238.1412487955
 7,2011-12-28 16:27:00,-3.0556238,5.170502,-7.855037,-273.6197747124566,2.84225861231486,13.720342424180773,25.930027431911892,303653.3159313615
 8,2011-12-28 16:28:00,-2.8697166,5.593325,-7.9218264,-273.6576911078559,1.992921193440755,13.691717359754774,25.793099933200413,304068.4906139276
-9,2011-12-28 16:29:00,-3.3140364,5.6189823,-7.282634,-273.6956075032552,1.1435837745666504,13.663092295328775,25.656172434488933,304483.6652964936
-10,2011-12-28 16:30:00,-3.4366038,6.0576234,-7.3536468,-273.73352389865454,0.294246355692545,13.634467230902779,25.51924493577745,304898.8399790597
+9,2011-12-28 16:29:00,-3.3140364,5.6189823,-7.282634,-273.6956075032552,1.1435837745666502,13.663092295328775,25.656172434488933,304483.6652964936
+10,2011-12-28 16:30:00,-3.4366038,6.0576234,-7.3536468,-273.73352389865454,0.2942463556925453,13.634467230902779,25.51924493577745,304898.8399790597
 11,2011-12-28 16:31:00,-3.381308,6.856737,-6.944566,-273.77144029405383,-0.5550910631815595,13.60584216647678,25.38231743706597,305314.0146616257
 12,2011-12-28 16:32:00,-3.328753,7.5127106,-5.5448656,-273.8093566894531,-1.404428482055664,13.577217102050781,25.245389938354492,305729.1893441918
 13,2011-12-28 16:33:00,-3.8617156,7.856515,-5.2628236,-273.60932540893555,-2.14258873462677,13.494311690330505,25.165552616119385,304874.7337442699
@@ -26,7 +26,7 @@
 24,2011-12-28 16:44:00,-1.6678288,7.6721487,2.0252714,-270.5620524088542,-8.744460105895996,13.379400783114964,24.594446182250977,311433.59706147993
 25,2011-12-28 16:45:00,-0.6572789,7.433451,3.2043815,-270.15028889973956,-9.103147506713867,13.49575752682156,24.591384887695312,314568.6101906458
 26,2011-12-28 16:46:00,-0.76850677,7.1484427,3.4156868,-269.738525390625,-9.461834907531738,13.612114270528158,24.58832359313965,317703.62331981165
-27,2011-12-28 16:47:00,-1.2839427,7.1281037,2.8385723,-269.32676188151044,-9.82052230834961,13.728471014234755,24.585262298583984,320838.6364489775
+27,2011-12-28 16:47:00,-1.2839427,7.1281037,2.8385723,-269.32676188151044,-9.82052230834961,13.728471014234755,24.585262298583984,320838.6364489776
 28,2011-12-28 16:48:00,-1.725095,6.9619064,2.3166437,-268.9149983723958,-10.17920970916748,13.844827757941353,24.58220100402832,323973.64957814344
 29,2011-12-28 16:49:00,-1.625965,6.968726,2.597164,-268.50323486328125,-10.537897109985352,13.96118450164795,24.579139709472656,327108.6627073093
 30,2011-12-28 16:50:00,-1.2849383,7.1886134,1.8534744,-268.65743679470484,-10.630468898349339,13.672580401102701,24.109635882907444,321899.9747955827
@@ -35,7 +35,7 @@
 33,2011-12-28 16:53:00,-0.8018301,7.907096,1.4127296,-269.1200425889757,-10.908184263441298,12.806768099466959,22.701124403211807,306273.91106040287
 34,2011-12-28 16:54:00,-0.8482224,7.8498273,1.616137,-269.2742445203993,-11.000756051805284,12.518163998921713,22.23162057664659,301065.2231486763
 35,2011-12-28 16:55:00,-1.2911748,7.251417,2.3932707,-269.42844645182294,-11.093327840169271,12.229559898376465,21.76211675008138,295856.5352369497
-36,2011-12-28 16:56:00,-1.4703364,7.087064,2.6657631,-269.5826483832465,-11.185899628533257,11.940955797831219,21.292612923516167,290647.84732522303
+36,2011-12-28 16:56:00,-1.4703364,7.087064,2.6657631,-269.5826483832465,-11.185899628533257,11.940955797831217,21.292612923516167,290647.84732522303
 37,2011-12-28 16:57:00,-1.428876,7.1809516,2.4042587,-269.73685031467016,-11.278471416897244,11.65235169728597,20.823109096950954,285439.15941349644
 38,2011-12-28 16:58:00,-1.3180223,7.2319584,1.9373779,-269.89105224609375,-11.37104320526123,11.363747596740723,20.353605270385742,280230.47150176985
 39,2011-12-28 16:59:00,-1.3039744,6.8620257,2.532516,-269.92820739746094,-10.234095275402069,11.66646683216095,20.103572368621826,283677.6268017735
