+++ output-data/GSE/Artemis/Artemis_2014-02-27_18-30.csv
@@ -4,11 +4,11 @@
 2,2014-02-27 18:32:00,6.835648,-20.002016,-3.920018,-438.4126281738281,-34.35730743408203,14.138419151306152,14.060135841369629,1029406.2472265492
 3,2014-02-27 18:33:00,6.602779,-19.2888,-4.3919487,-439.1054280598958,-26.232461293538414,11.050732824537489,14.455382029215494,1016216.5215990965
 4,2014-02-27 18:34:00,5.8409195,-15.992431,-10.298167,-439.79822794596356,-18.107615152994793,7.963046497768826,14.850628217061361,1003026.7959716436
-5,2014-02-27 18:35:00,4.3377085,-17.509218,-5.769117,-440.49102783203125,-9.982769012451172,4.875360171000162,15.245874404907227,989837.0703441909
+5,2014-02-27 18:35:00,4.3377085,-17.509218,-5.769117,-440.49102783203125,-9.982769012451174,4.875360171000162,15.245874404907227,989837.0703441909
 6,2014-02-27 18:36:00,5.328194,-14.334289,-11.692186,-441.18382771809894,-1.8579228719075545,1.7876738442314988,15.641120592753092,976647.3447167382
-7,2014-02-27 18:37:00,5.7076426,-13.735289,-12.518355,-441.8766276041667,6.266923268636063,-1.3000124825371646,16.036366780598957,963457.6190892854
-8,2014-02-27 18:38:00,5.273017,-12.571297,-13.612218,-442.5694274902344,14.391769409179688,-4.387698809305828,16.431612968444824,950267.8934618327
-9,2014-02-27 18:39:00,5.588456,-12.340416,-13.334734,-443.26222737630206,22.516615549723305,-7.475385136074493,16.82685915629069,937078.16783438
+7,2014-02-27 18:37:00,5.7076426,-13.735289,-12.518355,-441.8766276041667,6.266923268636065,-1.3000124825371646,16.036366780598957,963457.6190892854
+8,2014-02-27 18:38:00,5.273017,-12.571297,-13.612218,-442.5694274902344,14.391769409179684,-4.387698809305828,16.431612968444824,950267.8934618327
+9,2014-02-27 18:39:00,5.588456,-12.340416,-13.334734,-443.26222737630206,22.516615549723305,-7.475385136074491,16.82685915629069,937078.16783438
 10,2014-02-27 18:40:00,5.6627083,-10.87871,-16.395609,-443.9550272623698,30.641461690266922,-10.563071462843155,17.222105344136555,923888.4422069271
 11,2014-02-27 18:41:00,5.686751,-9.465114,-15.075491,-444.6478271484375,38.76630783081055,-13.650757789611816,17.617351531982422,910698.7165794744
 12,2014-02-27 18:42:00,4.3813367,-8.557464,-15.57814,-444.4168192545573,37.54148631625705,-13.655547036064995,17.394983927408855,913836.0980318501
@@ -44,14 +44,14 @@
 42,2014-02-27 19:12:00,-1.3673668,9.348044,-13.394591,-447.2056376139323,37.31055196126302,-15.012409740024143,18.22540834214952,891774.8353245093
 43,2014-02-27 19:13:00,-1.2928432,6.574082,-14.846842,-447.14381917317706,38.624071756998696,-14.865575260586208,18.5209772321913,884204.887364193
 44,2014-02-27 19:14:00,-0.9159351,5.70593,-14.160074,-447.0820007324219,39.937591552734375,-14.718740781148275,18.816546122233074,876634.9394038765
-45,2014-02-27 19:15:00,-0.8565199,0.88680965,-14.931593,-447.0201822916667,41.25111134847005,-14.57190630171034,19.11211501227485,869064.9914435602
+45,2014-02-27 19:15:00,-0.8565199,0.88680965,-14.931593,-447.0201822916667,41.251111348470054,-14.57190630171034,19.11211501227485,869064.9914435602
 46,2014-02-27 19:16:00,-0.96216094,2.31649,-14.992033,-446.95836385091144,42.56463114420573,-14.425071822272407,19.407683902316624,861495.0434832438
 47,2014-02-27 19:17:00,-1.4700159,4.595142,-14.126794,-446.89654541015625,43.878150939941406,-14.278237342834473,19.7032527923584,853925.0955229275
 48,2014-02-27 19:18:00,1.0810052,4.667414,-13.622321,-448.14312744140625,44.84344139099121,-13.271713352203369,19.54001636505127,853922.4128988804
 49,2014-02-27 19:19:00,-1.4634293,8.007974,-13.220907,-449.38970947265625,45.80873184204101,-12.265189361572265,19.37677993774414,853919.7302748335
 50,2014-02-27 19:20:00,-0.19045122,11.267343,-12.008257,-450.63629150390625,46.77402229309082,-11.258665370941163,19.213543510437013,853917.0476507865
 51,2014-02-27 19:21:00,-4.7760572,5.7020526,-13.2132435,-451.88287353515625,47.739312744140626,-10.252141380310059,19.050307083129884,853914.3650267394
-52,2014-02-27 19:22:00,-2.8339264,8.319551,-12.624529,-453.12945556640625,48.70460319519043,-9.245617389678955,18.887070655822754,853911.6824026925
+52,2014-02-27 19:22:00,-2.8339264,8.319551,-12.624529,-453.12945556640625,48.70460319519043,-9.245617389678955,18.887070655822754,853911.6824026924
 53,2014-02-27 19:23:00,-3.3875368,9.106469,-11.525508,-454.37603759765625,49.66989364624023,-8.239093399047853,18.723834228515624,853908.9997786455
 54,2014-02-27 19:24:00,-2.4825363,9.27073,-11.802169,-455.62261962890625,50.635184097290036,-7.232569408416749,18.560597801208495,853906.3171545984
 55,2014-02-27 19:25:00,-2.3803585,8.767374,-11.565535,-456.86920166015625,51.60047454833985,-6.226045417785645,18.39736137390137,853903.6345305514
@@ -128,7 +128,7 @@
 126,2014-02-27 20:36:00,3.8879783,-0.0041691405,-12.573395,-448.7643676757813,44.25625076293945,-24.54074897766113,17.461295700073244,742482.0399667815
 127,2014-02-27 20:37:00,3.6767578,-0.31641027,-12.303856,-447.5056182861328,41.8467529296875,-24.836880493164063,17.729839515686034,748027.5329507593
 128,2014-02-27 20:38:00,5.0879426,-2.276,-12.559783,-446.2468688964844,39.43725509643555,-25.13301200866699,17.99838333129883,753573.0259347371
-129,2014-02-27 20:39:00,4.756954,-1.4633892,-12.591494,-444.9881195068359,37.02775726318359,-25.429143524169923,18.26692714691162,759118.518918715
+129,2014-02-27 20:39:00,4.756954,-1.4633892,-12.591494,-444.9881195068359,37.027757263183595,-25.429143524169923,18.26692714691162,759118.518918715
 130,2014-02-27 20:40:00,0.84236544,1.2128621,-11.757586,-443.7293701171875,34.61825942993164,-25.72527503967285,18.535470962524414,764664.0119026927
 131,2014-02-27 20:41:00,0.8773133,1.291271,-12.22445,-444.5772705078125,35.78374693128798,-25.0946716732449,18.143400722079807,767271.8431714444
 132,2014-02-27 20:42:00,0.35925773,1.9972229,-10.857063,-445.4251708984375,36.94923443264432,-24.464068306816948,17.7513304816352,769879.6744401961
@@ -145,7 +145,7 @@
 143,2014-02-27 20:53:00,-2.8288913,-1.4687217,-11.176625,-445.72820366753473,42.544724782307945,-23.000815709431965,16.368141916063095,753317.9681001074
 144,2014-02-27 20:54:00,1.178374,-2.1789222,-10.626317,-444.32013617621527,41.90399424235026,-23.738558451334637,16.708467695448135,744613.8367947699
 145,2014-02-27 20:55:00,2.6769373,-2.2752595,-10.211983,-442.9120686848958,41.26326370239258,-24.476301193237305,17.04879347483317,735909.7054894322
-146,2014-02-27 20:56:00,3.2959979,-2.4192657,-9.674913,-441.5040011935764,40.6225331624349,-25.214043935139976,17.38911925421821,727205.5741840946
+146,2014-02-27 20:56:00,3.2959979,-2.4192657,-9.674913,-441.5040011935764,40.6225331624349,-25.214043935139973,17.38911925421821,727205.5741840946
 147,2014-02-27 20:57:00,4.36292,-2.523079,-10.555708,-440.09593370225696,39.98180262247721,-25.951786677042644,17.729445033603245,718501.442878757
 148,2014-02-27 20:58:00,6.0582147,-2.0170224,-10.219543,-438.6878662109375,39.34107208251953,-26.689529418945312,18.06977081298828,709797.3115734194
 149,2014-02-27 20:59:00,5.659242,-3.360606,-10.215431,-439.4786885579427,37.06601100497775,-26.63294219970703,17.898840586344402,719311.6209178591
