+++ output-data/GSE/Artemis/Artemis_2014-04-05_19-30.csv
@@ -56,7 +56,7 @@
 54,2014-04-05 20:24:00,4.739501,5.6808057,-5.9915004,-430.1359354654948,66.36179436577692,-45.133250342475044,14.33032512664795,450702.42287119775
 55,2014-04-05 20:25:00,4.687161,5.826238,-6.131508,-429.64027913411456,66.67082892523871,-44.787430657280815,14.584074974060059,447894.07357068185
 56,2014-04-05 20:26:00,3.2190475,5.5958047,-6.8297586,-429.1446228027344,66.97986348470052,-44.441610972086586,14.837824821472168,445085.72427016596
-57,2014-04-05 20:27:00,3.6092792,5.251738,-7.0697474,-428.6489664713542,67.28889804416232,-44.095791286892364,15.091574668884277,442277.37496965006
+57,2014-04-05 20:27:00,3.6092792,5.251738,-7.0697474,-428.6489664713542,67.28889804416232,-44.095791286892364,15.091574668884277,442277.37496965
 58,2014-04-05 20:28:00,3.6073186,3.882169,-7.4315333,-428.15331013997394,67.59793260362413,-43.749971601698135,15.345324516296387,439469.0256691341
 59,2014-04-05 20:29:00,2.9733262,3.4620888,-7.920659,-427.65765380859375,67.90696716308594,-43.404151916503906,15.599074363708496,436660.6763686182
 60,2014-04-05 20:30:00,2.7286103,3.1405342,-7.660374,-427.7433369954427,67.90199449327257,-43.38200887044271,15.411814159817165,440743.45899962704
@@ -200,7 +200,7 @@
 198,2014-04-05 22:48:00,6.072619,5.782258,0.36458915,-417.56051296657984,66.9866188897027,-40.539758894178604,18.699662950303818,482301.2108846611
 199,2014-04-05 22:49:00,4.7215686,4.9881477,-0.7322105,-418.28968302408856,68.10766347249348,-41.021870930989586,19.148251215616863,467584.17501526594
 200,2014-04-05 22:50:00,2.9832435,3.8621411,-2.7955077,-419.01885308159723,69.22870805528429,-41.50398296780057,19.596839480929905,452867.1391458707
-201,2014-04-05 22:51:00,3.2517755,3.0333073,-0.5411925,-419.7480231391059,70.34975263807509,-41.98609500461154,20.045427746242947,438150.10327647557
+201,2014-04-05 22:51:00,3.2517755,3.0333073,-0.5411925,-419.7480231391059,70.34975263807509,-41.98609500461154,20.045427746242947,438150.1032764755
 202,2014-04-05 22:52:00,6.450368,4.9031878,4.9263353,-420.47719319661456,71.47079722086589,-42.468207041422524,20.49401601155599,423433.0674070803
 203,2014-04-05 22:53:00,5.916075,3.3467634,7.4757,-421.2063632541233,72.59184180365668,-42.950319078233505,20.942604276869034,408716.03153768514
 204,2014-04-05 22:54:00,5.6256633,1.2996761,9.223066,-421.93553331163196,73.71288638644748,-43.43243111504449,21.391192542182075,393998.99566828995
