+++ output-data/GSE/Artemis/Artemis_2015-03-16_03-50.csv
@@ -94,7 +94,7 @@
 92,2015-03-16 05:22:00,4.427084,-2.6487021,2.3900526,-331.13795776367186,55.03184471130371,-11.4228009223938,17.519674873352052,282740.1725402785
 93,2015-03-16 05:23:00,4.345857,-2.613961,2.3735254,-330.7575622558594,54.89188079833984,-11.333251190185546,17.423172760009766,291484.5928845443
 94,2015-03-16 05:24:00,4.3604083,-3.0291462,2.5113432,-330.3771667480469,54.75191688537598,-11.243701457977295,17.32667064666748,300229.01322881004
-95,2015-03-16 05:25:00,4.477336,-2.58803,2.7406619,-329.99677124023435,54.61195297241211,-11.154151725769044,17.230168533325195,308973.4335730758
+95,2015-03-16 05:25:00,4.477336,-2.58803,2.7406619,-329.99677124023435,54.61195297241211,-11.154151725769044,17.230168533325195,308973.43357307574
 96,2015-03-16 05:26:00,3.9044306,-2.4509523,2.2054849,-329.6163757324219,54.47198905944824,-11.06460199356079,17.13366641998291,317717.8539173415
 97,2015-03-16 05:27:00,3.525055,-2.6086788,1.0866072,-329.23598022460936,54.332025146484376,-10.97505226135254,17.037164306640626,326462.27426160726
 98,2015-03-16 05:28:00,3.330013,-2.0915735,0.5974812,-328.8555847167969,54.192061233520505,-10.885502529144286,16.94066219329834,335206.694605873
