+++ output-data/GSE/Artemis/Artemis_2015-06-12_17-10.csv
@@ -21,7 +21,7 @@
 19,2015-06-12 17:29:00,2.4945145,2.2429323,-1.865516,-430.6709391276042,-31.928399403889973,-74.77705891927083,9.690203666687012,807774.9017870413
 20,2015-06-12 17:30:00,2.5149212,2.6080818,-1.5981331,-431.13232760959204,-30.474085913764107,-74.33398098415799,9.589940388997396,812079.0427103336
 21,2015-06-12 17:31:00,2.6326265,2.8237827,-1.700507,-431.59371609157984,-29.019772423638237,-73.89090304904514,9.48967711130778,816383.1836336259
-22,2015-06-12 17:32:00,2.8824174,2.99909,-1.5723622,-432.0551045735677,-27.565458933512367,-73.4478251139323,9.389413833618164,820687.3245569181
+22,2015-06-12 17:32:00,2.8824174,2.99909,-1.5723622,-432.0551045735677,-27.56545893351237,-73.4478251139323,9.389413833618164,820687.3245569181
 23,2015-06-12 17:33:00,2.987671,2.8728843,-1.5551854,-432.51649305555554,-26.1111454433865,-73.00474717881944,9.289150555928549,824991.4654802103
 24,2015-06-12 17:34:00,3.0085118,3.009557,-1.3912537,-432.9778815375434,-24.656831953260635,-72.5616692437066,9.188887278238932,829295.6064035025
 25,2015-06-12 17:35:00,3.3209472,2.654646,-1.0292119,-433.43927001953125,-23.202518463134766,-72.11859130859375,9.088624000549316,833599.7473267948
@@ -57,7 +57,7 @@
 55,2015-06-12 18:05:00,1.6809905,2.1062608,-1.0884726,-431.7696940104167,-13.255353291829428,-57.057865142822266,9.995218276977539,899543.250245399
 56,2015-06-12 18:06:00,1.5655823,1.7941258,-0.9886904,-432.1029798719618,-13.995928234524197,-57.50198745727539,9.993796666463217,898501.1585212653
 57,2015-06-12 18:07:00,1.4914682,1.2407398,-1.2185348,-432.43626573350696,-14.736503177218967,-57.946109771728516,9.992375055948893,897459.0667971314
-58,2015-06-12 18:08:00,1.1876154,2.323091,-1.0807784,-432.76955159505206,-15.477078119913738,-58.39023208618164,9.99095344543457,896416.9750729977
+58,2015-06-12 18:08:00,1.1876154,2.323091,-1.0807784,-432.76955159505206,-15.477078119913736,-58.39023208618164,9.99095344543457,896416.9750729977
 59,2015-06-12 18:09:00,0.7743597,2.8945148,-0.3506967,-433.10283745659723,-16.217653062608505,-58.834354400634766,9.989531834920248,895374.883348864
 60,2015-06-12 18:10:00,0.43308082,2.9951992,0.1218727,-433.43612331814234,-16.958228005303276,-59.27847671508789,9.988110224405924,894332.7916247303
 61,2015-06-12 18:11:00,-0.42742556,1.8220531,0.054481946,-433.7694091796875,-17.698802947998047,-59.722599029541016,9.986688613891602,893290.6999005966
@@ -84,9 +84,9 @@
 82,2015-06-12 18:32:00,0.9133021,2.7266643,-1.2643572,-430.08126491970484,8.134230931599934,-55.0962036980523,10.137583520677355,896462.6396891555
 83,2015-06-12 18:33:00,-0.46258217,4.009485,0.18370111,-429.85120646158856,5.237308502197266,-55.09979756673177,9.96700922648112,896577.6373976611
 84,2015-06-12 18:34:00,-0.31688341,2.7863212,-0.00568073,-429.62114800347223,2.3403860727945958,-55.10339143541124,9.796434932284885,896692.6351061667
-85,2015-06-12 18:35:00,-0.1694931,3.0020661,-0.06407632,-429.3910895453559,-0.5565363566080741,-55.10698530409071,9.62586063808865,896807.6328146722
-86,2015-06-12 18:36:00,-0.3219767,3.9098911,-0.27593455,-429.16103108723956,-3.453458786010742,-55.11057917277018,9.455286343892416,896922.6305231778
-87,2015-06-12 18:37:00,-0.25003842,3.7666376,-0.31478062,-428.9309726291233,-6.350381215413414,-55.114173041449654,9.28471204969618,897037.6282316834
+85,2015-06-12 18:35:00,-0.1694931,3.0020661,-0.06407632,-429.3910895453559,-0.5565363566080737,-55.10698530409071,9.62586063808865,896807.6328146722
+86,2015-06-12 18:36:00,-0.3219767,3.9098911,-0.27593455,-429.16103108723956,-3.453458786010743,-55.11057917277018,9.455286343892416,896922.6305231778
+87,2015-06-12 18:37:00,-0.25003842,3.7666376,-0.31478062,-428.9309726291233,-6.350381215413412,-55.114173041449654,9.28471204969618,897037.6282316834
 88,2015-06-12 18:38:00,-0.28912476,3.8899944,-0.3596517,-428.70091417100696,-9.247303644816082,-55.11776691012912,9.114137755499947,897152.625940189
 89,2015-06-12 18:39:00,-0.35106152,4.4105697,-0.41783166,-428.4708557128906,-12.14422607421875,-55.121360778808594,8.943563461303711,897267.6236486946
 90,2015-06-12 18:40:00,-0.4481583,3.9527707,-0.57060057,-428.4088406032986,-10.929211258888245,-55.517520904541016,8.930834346347385,894655.9263320864
