+++ output-data/GSE/Artemis/Artemis_2016-08-02_12-15.csv
@@ -8,7 +8,7 @@
 6,2016-08-02 12:21:00,6.519169,1.9058734,1.2087065,-346.1501057942708,57.35292053222656,34.676228841145836,16.07724444071452,207049.2105389889
 7,2016-08-02 12:22:00,7.334796,1.7211198,0.55933577,-347.6715359157986,57.49318186442057,34.93935309516059,15.470874362521702,207868.7585795753
 8,2016-08-02 12:23:00,7.470666,2.062224,1.0085946,-349.1929660373264,57.633443196614586,35.202477349175346,14.864504284328884,208688.30662016175
-9,2016-08-02 12:24:00,7.2195897,2.8723996,1.7594053,-350.7143961588542,57.773704528808594,35.4656016031901,14.258134206136067,209507.85466074815
+9,2016-08-02 12:24:00,7.2195897,2.8723996,1.7594053,-350.7143961588542,57.773704528808594,35.4656016031901,14.258134206136068,209507.85466074815
 10,2016-08-02 12:25:00,8.3754015,3.4671874,2.3572783,-352.23582628038196,57.9139658610026,35.728725857204864,13.65176412794325,210327.40270133456
 11,2016-08-02 12:26:00,8.92435,4.0239344,2.0283098,-353.75725640190973,58.05422719319662,35.99185011121962,13.045394049750435,211146.950741921
 12,2016-08-02 12:27:00,8.742789,3.8948073,3.0438855,-355.2786865234375,58.194488525390625,36.254974365234375,12.439023971557617,211966.4987825074
@@ -17,7 +17,7 @@
 15,2016-08-02 12:30:00,9.7234745,4.0642457,3.8577332,-359.4144714355469,58.59677352905273,36.1942943572998,12.497668933868407,224184.51000458654
 16,2016-08-02 12:31:00,10.314667,4.2792993,3.2127175,-360.79306640625,58.73086853027344,36.174067687988284,12.517217254638672,228257.18041194623
 17,2016-08-02 12:32:00,9.213148,3.4948175,5.96868,-362.1716613769531,58.86496353149414,36.15384101867676,12.536765575408936,232329.85081930595
-18,2016-08-02 12:33:00,8.540542,3.5082529,6.980977,-363.5502563476563,58.99905853271484,36.13361434936523,12.5563138961792,236402.52122666565
+18,2016-08-02 12:33:00,8.540542,3.5082529,6.980977,-363.5502563476563,58.99905853271484,36.13361434936523,12.5563138961792,236402.52122666567
 19,2016-08-02 12:34:00,7.36534,2.652465,8.8102,-364.92885131835936,59.13315353393555,36.11338768005371,12.575862216949464,240475.19163402537
 20,2016-08-02 12:35:00,7.336277,2.7057087,8.470514,-366.3074462890625,59.26724853515625,36.093161010742186,12.595410537719726,244547.8620413851
 21,2016-08-02 12:36:00,6.0790377,2.3035712,9.12265,-367.6860412597656,59.401343536376956,36.07293434143067,12.61495885848999,248620.53244874478
@@ -137,7 +137,7 @@
 135,2016-08-02 14:30:00,11.163947,7.706707,4.8035483,-369.18967692057294,57.79361343383789,34.00970204671224,16.755015055338543,276704.5694550399
 136,2016-08-02 14:31:00,11.3405695,8.392811,4.3046994,-369.69021945529516,57.23788579305013,33.669547186957466,17.148270818922256,284583.2134668818
 137,2016-08-02 14:32:00,11.235306,7.9152093,4.2937627,-370.19076199001734,56.68215815226237,33.32939232720269,17.541526582505966,292461.8574787236
-138,2016-08-02 14:33:00,11.189608,7.9243035,2.3898804,-370.69130452473956,56.12643051147461,32.989237467447914,17.93478234608968,300340.5014905655
+138,2016-08-02 14:33:00,11.189608,7.9243035,2.3898804,-370.69130452473956,56.12643051147461,32.989237467447914,17.93478234608968,300340.5014905654
 139,2016-08-02 14:34:00,10.750999,8.351964,2.3667061,-371.1918470594618,55.57070287068685,32.649082607693146,18.328038109673393,308219.1455024073
 140,2016-08-02 14:35:00,9.839635,8.284139,1.9101483,-371.692389594184,55.014975229899086,32.30892774793837,18.721293873257107,316097.78951424913
 141,2016-08-02 14:36:00,8.21586,6.641698,0.5774352,-372.19293212890625,54.45924758911133,31.968772888183594,19.11454963684082,323976.433526091
@@ -176,7 +176,7 @@
 174,2016-08-02 15:09:00,-3.868145,-10.55165,12.990358,-393.5115600585938,62.995453643798825,36.98898849487305,9.667409515380859,342014.8535756975
 175,2016-08-02 15:10:00,-0.7988059,-8.191535,15.634999,-395.0029327392578,63.18994674682617,37.44093818664551,9.205488395690917,345526.5434697824
 176,2016-08-02 15:11:00,0.46333367,-6.1678276,16.619638,-396.4943054199219,63.384439849853514,37.892887878417966,8.743567276000977,349038.23336386733
-177,2016-08-02 15:12:00,0.99937916,-5.614602,16.83031,-397.9856781005859,63.57893295288086,38.34483757019043,8.281646156311034,352549.9232579522
+177,2016-08-02 15:12:00,0.99937916,-5.614602,16.83031,-397.9856781005859,63.57893295288086,38.34483757019043,8.281646156311036,352549.9232579522
 178,2016-08-02 15:13:00,0.7099914,-4.9675727,17.088564,-399.47705078125,63.7734260559082,38.79678726196289,7.819725036621094,356061.6131520371
 179,2016-08-02 15:14:00,0.41136017,-4.4455137,17.297804,-399.47705078125,63.7734260559082,38.79678726196289,7.819725036621094,356061.6131520371
 180,2016-08-02 15:15:00,1.392445,-4.0460877,17.40035,-399.47705078125,63.7734260559082,38.79678726196289,7.819725036621094,356061.6131520371
