+++ output-data/GSE/Artemis/Artemis_2018-10-09_04-00.csv
@@ -29,14 +29,14 @@
 27,2018-10-09 04:27:00,-4.88261,1.1324084,-0.59682167,-565.9449869791666,85.15933736165364,-7.48110564549764,2.5252823299831815,900456.8573612527
 28,2018-10-09 04:28:00,-5.126615,0.9268833,-0.5390407,-565.9502766927084,85.38489786783855,-7.4514390627543134,2.5399339728885226,879249.3222957134
 29,2018-10-09 04:29:00,-5.049975,0.8974105,-0.3923549,-565.95556640625,85.61045837402344,-7.421772480010986,2.554585615793864,858041.7872301739
-30,2018-10-09 04:30:00,-5.023675,0.7809731,-0.4573155,-565.9608561197916,85.83601888020833,-7.392105897267659,2.5692372586992054,836834.2521646346
+30,2018-10-09 04:30:00,-5.023675,0.7809731,-0.4573155,-565.9608561197916,85.83601888020833,-7.392105897267659,2.5692372586992054,836834.2521646345
 31,2018-10-09 04:31:00,-4.9189,1.0776483,-1.1919569,-565.9661458333334,86.06157938639323,-7.362439314524333,2.5838889016045465,815626.7170990952
 32,2018-10-09 04:32:00,-4.9355955,1.3451779,-1.4029653,-565.971435546875,86.28713989257812,-7.332772731781006,2.5985405445098877,794419.1820335558
 33,2018-10-09 04:33:00,-4.732576,1.6070689,-1.5494177,-565.9227227105034,85.87681918674045,-7.3966941303677025,2.6129292382134333,808453.41777904
 34,2018-10-09 04:34:00,-4.659997,1.4685456,-1.7592252,-565.874009874132,85.46649848090277,-7.4606155289544,2.627317931916979,822487.6535245243
 35,2018-10-09 04:35:00,-4.4508543,1.6733788,-2.1623218,-565.8252970377604,85.05617777506511,-7.524536927541097,2.641706625620524,836521.8892700085
 36,2018-10-09 04:36:00,-4.330758,1.7309843,-2.230775,-565.7765842013889,84.64585706922743,-7.588458326127794,2.6560953193240695,850556.1250154928
-37,2018-10-09 04:37:00,-4.3530545,1.8611181,-1.9722744,-565.7278713650173,84.23553636338976,-7.652379724714491,2.670484013027615,864590.360760977
+37,2018-10-09 04:37:00,-4.3530545,1.8611181,-1.9722744,-565.7278713650173,84.23553636338976,-7.652379724714491,2.670484013027615,864590.3607609771
 38,2018-10-09 04:38:00,-4.303632,1.8870312,-2.0778723,-565.6791585286459,83.82521565755208,-7.7163011233011884,2.6848727067311606,878624.5965064614
 39,2018-10-09 04:39:00,-4.502434,1.8223572,-1.902487,-565.6304456922743,83.41489495171442,-7.780222521887885,2.6992614004347057,892658.8322519456
 40,2018-10-09 04:40:00,-4.3844113,1.9345869,-1.7660531,-565.5817328559028,83.00457424587674,-7.844143920474583,2.7136500941382513,906693.0679974299
@@ -63,7 +63,7 @@
 61,2018-10-09 05:01:00,-1.3238195,3.6904051,2.0402749,-583.7946899414062,84.66834106445313,-10.507378768920898,2.54461612701416,1019814.7949796005
 62,2018-10-09 05:02:00,-2.6634846,3.0228333,1.9091791,-583.7163269042969,84.87173233032226,-10.574560928344727,2.565183734893799,1005684.289723853
 63,2018-10-09 05:03:00,-2.9755797,2.4817407,2.3213933,-583.6379638671875,85.0751235961914,-10.641743087768555,2.5857513427734373,991553.7844681055
-64,2018-10-09 05:04:00,-3.0653899,2.0625632,2.4884343,-583.5596008300781,85.27851486206055,-10.708925247192383,2.606318950653076,977423.279212358
+64,2018-10-09 05:04:00,-3.0653899,2.0625632,2.4884343,-583.5596008300781,85.27851486206055,-10.708925247192383,2.606318950653076,977423.2792123581
 65,2018-10-09 05:05:00,-3.1118882,2.840493,1.5381511,-583.4812377929687,85.48190612792969,-10.77610740661621,2.626886558532715,963292.7739566105
 66,2018-10-09 05:06:00,-3.0491605,3.1636739,1.080149,-583.4028747558593,85.68529739379883,-10.843289566040038,2.6474541664123534,949162.268700863
 67,2018-10-09 05:07:00,-2.430122,3.798084,0.6666411,-583.32451171875,85.88868865966796,-10.910471725463868,2.6680217742919923,935031.7634451155
@@ -147,8 +147,8 @@
 145,2018-10-09 06:25:00,-2.9797788,2.5544403,-2.431795,-578.8206380208334,77.38143920898438,-8.551594416300455,2.4086598555246987,1050606.7092935927
 146,2018-10-09 06:26:00,-3.3351283,2.3645465,-2.5441093,-580.4686957465278,76.1342290242513,-8.837654325697157,2.287367476357354,1094380.2083012788
 147,2018-10-09 06:27:00,-3.113501,3.2194517,-2.1832936,-582.1167534722222,74.88701883951823,-9.123714235093859,2.166075097190009,1138153.7073089646
-148,2018-10-09 06:28:00,-2.295691,3.9415374,-1.7522093,-583.7648111979166,73.63980865478516,-9.40977414449056,2.044782718022664,1181927.2063166504
-149,2018-10-09 06:29:00,-0.4978889,5.0424743,-0.42040396,-585.4128689236111,72.39259847005208,-9.69583405388726,1.9234903388553195,1225700.7053243364
+148,2018-10-09 06:28:00,-2.295691,3.9415374,-1.7522093,-583.7648111979166,73.63980865478516,-9.40977414449056,2.0447827180226645,1181927.2063166504
+149,2018-10-09 06:29:00,-0.4978889,5.0424743,-0.42040396,-585.4128689236111,72.39259847005208,-9.69583405388726,1.9234903388553195,1225700.7053243362
 150,2018-10-09 06:30:00,0.1229527,5.0897193,-0.4951445,-587.0609266493055,71.14538828531902,-9.981893963283962,1.8021979596879747,1269474.2043320222
 151,2018-10-09 06:31:00,-0.060433142,5.126477,0.013426123,-588.708984375,69.89817810058594,-10.267953872680664,1.6809055805206299,1313247.703339708
 152,2018-10-09 06:32:00,-0.78380257,4.6831484,0.12102611,-587.2179931640625,70.15832748413087,-10.13128719329834,1.7222461223602294,1284865.0185632692
