+++ output-data/GSE/Artemis/Artemis_2019-11-27_13-00.csv
@@ -37,7 +37,7 @@
 35,2019-11-27 13:35:00,-0.88922805,4.216856,-0.7903238,-375.2042541503906,22.06638018290202,7.505797704060873,4.812869548797607,131977.33043813077
 36,2019-11-27 13:36:00,-1.0386039,4.1214075,-1.1940318,-376.1011047363281,21.1775025261773,8.197495248582628,4.8003471692403155,131664.7289896612
 37,2019-11-27 13:37:00,-2.1037385,3.8399522,-0.7183737,-376.9979553222656,20.288624869452583,8.889192793104383,4.787824789683024,131352.12754119164
-38,2019-11-27 13:38:00,-1.7940171,4.040071,-0.82908875,-377.8948059082031,19.399747212727867,9.58089033762614,4.775302410125732,131039.52609272205
+38,2019-11-27 13:38:00,-1.7940171,4.040071,-0.82908875,-377.8948059082031,19.399747212727863,9.58089033762614,4.775302410125732,131039.52609272205
 39,2019-11-27 13:39:00,-1.8048575,3.9597104,-1.0428966,-378.7916564941406,18.510869556003147,10.272587882147896,4.7627800305684405,130726.92464425247
 40,2019-11-27 13:40:00,-1.8552464,3.9197488,-1.1700188,-379.6885070800781,17.621991899278427,10.964285426669651,4.750257651011149,130414.3231957829
 41,2019-11-27 13:41:00,-2.1837866,3.8643591,-0.72648346,-380.5853576660156,16.73311424255371,11.655982971191406,4.737735271453857,130101.72174731332
